+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-## 1.0.1
+## 1.1.0
 
 - Change log and add verbose option.
 - Display runtime after download ends.
+- Add download speed.
+- Display merge progress.
+- Support ext name
 
 ## 1.0.0